+++ quarkc/_metadata.py
@@ -20,7 +20,7 @@ __all__ = [
 ]
 
 __title__ = 'datawire-quarkdev'
-__version__ = '1.0.332'
+__version__ = '1.0.376'
 
 __summary__ = "Quark: an IDL for high level (micro)service interfaces"
 __uri__ = "http://datawire.github.io/quark/"